Refactor "Service" to "Bearer"
* More accurately reflects RadioDNS terminology
* Previous module/class names mapped to ensure backwards compatability
* Unit tests introduce for new naming and previous unit tests retained in
  `legacy` to maintain regression testing
