+++ pyradiodns/radiodns/__init__.py
@@ -1,0 +1,1 @@
+# maps the legacy module/class name to the new one
