+++ test/legacy/test_dab_service.py
@@ -1,6 +1,8 @@
 import unittest
 from pyradiodns.radiodns.dab_service import RadioDNS_DABService
 
+# this tests backwards compatability with the former 'Service' classes
+
 class TestRadioDNSDABService(unittest.TestCase):
 
     def setUp(self):
