+++ test/legacy/test_hd_service.py
@@ -1,6 +1,8 @@
 import unittest
 from pyradiodns.radiodns.hd_service import RadioDNS_HDService
 
+# this tests backwards compatability with the former 'Service' classes
+
 class TestRadioDNSHDService(unittest.TestCase):
 
     def setUp(self):
